[RF] New function for server changing without setting attributes
If one wants to replace some servers of a RooAbsArg with
other servers, this is the the only option you have:

```c++
bool RooAbsArg::redirectServers(const RooAbsCollection& newSetOrig,
                                bool mustReplaceAll, bool nameChange,
                                bool isRecursionStep)
```

That's not so great, because if the new servers have different names,
thes old names must be set as the `"ORIGNAME:<myname>"` attribute for
the function to know which servers should be replaced.

These attributes are generally not reset, and this mutation of the
server attributes can cause trouble down the line.

Notably, this causes trouble in the BatchMode when compiling the
computation graph for a given normalization set. One single new server
might replace differently-named original servers. In that case, the new
servers get *two* "ORIGNAME:" attributes and the the `redirectServers()`
function doesn't know what to do.

Also, it's inefficient because of the string matching.

For this reason, a new overload is suggested, where one can simply
specify what should replace what in a lookup map, and it doesn't matter
if the names are different:

```c++
bool RooAbsArg::redirectServers(
    std::unordered_map<RooAbsArg *, RooAbsArg *> const &replacements
)
```

The implementation of `redirectServers()` calls
`RooAbsProxy::changePointer()`, for which such a new overload also had
to be added.