[RF] Better server redirection in compileForNormSet without attributes
Using the new `RooAbsArg::redirectServers()` overload that takes a
lookup map of server replacements, the NormalizationHelpers can be
improved such that it's not necessary to set the `"ORIGNAME:"`
attributes of new servers.

This fixes some failures in the RooFit tutorials with BatchMode where
some servers eventually aggregated multiple `"ORIGNAME:"` attributes,
causing `redirectServers()` to error out.