[math][minuit2] Fix in NegativeG2LineSearch for external gradient
When an external gradient and Hessian is provided to Minuit2 in NG2LS when called from MnSeed  the full Hessian was returned. This is not helpful and causing
the minimization to fail immediatly.

It is better to just use the diagobal G2 also if the Hessian can be computed.
It is also found to use in this case for the line search a  dummy step size = 1.

Improve also some print message in other Minuit2 classes and classify the messages from N2PG callulator and LS as Trace messages