[ntuple] Make sure that all tasks have finished in `~RPageStorage`
In the `RPageStorage` destructor: wait for unterminated tasks, if any,
as they may still hold a reference to `this`.

This is known to happen, e.g. in `RPageSinkBuf` if `CommitClusterImpl()`
is not called after committing some pages via `CommitPageImpl()` and
the `RPageSinkBuf` is immediately destructed.

This fixes the ntuple_extended:`RNTuple.SmallClusters` test.