[RF] Forbid resetting RooFormula
The `RooFormula` class allows resetting of the fomula, but this should
better not be done because it greatly increases the error surface.

The only reason why this was needed anyway was the implementation of
`RooGenericPdf::readFromStream()`, but this functionality is not
required. Reading from input streams was never completely implemented in
RooFit anyway, the only place where it's used is the I/O of RooDataSet,
and these don't use the `RooGenericPdf`.