[RF] Add exception if `RooAbsArg::_dataToken` is corrupted
The RooFitDriver mutates the RooAbsArgs it evaluates by setting a
specific index token to quickly look up cached information for a given
RooAbsArg.

Because of this, it's not possible to evaluate two computation graphs
that share RooAbsArgs with two separate RooFitDrivers.

So far, this concurrent evaluation would silently give wrong results,
and this commit suggests to throw an exception if the data tokens are
set by different RooFitDrivers before they are reset by the first one.