[RF] Support multiple NLL with BatchMode existing at the same time
As explained in the last commit, it's not possible to evaluate two
different likelihoods that share RooAbsArgs by different RooFitDrivers.

However, this usecase needs to be supported, because in RooStats is
often happens that different likelihoods based on the same model are
instantiated.

The key to the solution in this commit is the insight that for each
likelihood, almost the full compute graph is cloned anyway, *except* for
the parameters. Therefore, the solution proposed in this commit is to
not set the data tokens for the parameters and not fill the data map
with their values. Instead, when querying their values, they are just
retrieved from the object itself. This is possible because parameters
are always scalar, and the RooAbsReal can already cache scalar members
in the `_value` field.