[RF] Add missing declarations for JSONInterface template specializations
In the RooFit `JSONInterface` there are some template specializations
that are defined in the translation unit but are not declared in the
header file.

This causes linker errors, because the compiler doesn't know that it has
to look for the specializations, as explained here:
https://stackoverflow.com/questions/11773960/do-template-specialisations-belong-into-the-header-or-source-file

To solve this problem, these small one-line functions are just inlined
in the header.