[DF] A mechanism to clone actions and their results
Actions of an RDataFrame computation graph can now be cloned, i.e. a new
`RAction` object is created with the same type of action helper and a
new result that will be filled when the cloned action is executed.
Cloning an action exploits the already present registering/deregistering
mechanism from `RLoopManager` happening after construction/destruction
of an `RAction`. This mechanism enables workflows that avoid re-jitting
the same nodes of the computation graph in case the same type of action
needs to be re-executed (for example, in the context of different tasks
being run in a process that refer to the same computation graph).