[cmake] Refactor `relatedrepo_Clone`
```
relaterepo_Checkout(REPO_NAME <repo> FETCHURL <url> FETCHREF <ref>
                    [FORCE] [REPO_DIR_VARIABLE <repo_dir>])
```
Clone/fetch from remote and checkout the specified commit-ish from the given
repository.  If a local clone of the repository is found, refuse to update it
unless the `FORCE` option is in place.
Given that the function may update an existing repository, the function has
been renamed to `relatedrepo_Checkout`.