[ntuple] Fix a memory leak in `RPageSinkBuf`
RPageSinkBuf calls `ReservePage()` as part of a `CommitPageImpl()` call.
This memory is release in `CommitClusterImpl()` by calling `ReleasePage()`
accordingly.

However, if `CommitClusterImpl()` is never invoked (e.g., because the
'small cluster' check failed in RNTuple.cxx:336), this memory is
leaked.
This patch fixes that by making `RColumnBuf` adopt the pages.
`DrainBufferedPages()` gives back the ownership to the caller, who is then
responsible of calling `ReleasePage()`.