[imt] Introduce env var `ROOT_MAX_THREADS`:
This overrides the number of threads that ROOT will be using in the thread pool.
This can be set to be higher or lower than the number of physical cores; it can
be specified as decimal, octal, or hex.

Fixes issue #9805.

Mention the env var in `ROOT::EnableImplicitMT()`'s doc.