[Math][Minuit2] New Minuit2 strategy for improved Hessian calculation
New Minuit2 strategy for improved Hessian calculation and return without
making positive definite.

This proposed new Strategy in minuit2 is the same migrad behaviour as
Strategy=2 but with the following changes to the Hesse calculation:

  * The step and g2 tolerances have been zeroed so that the 7 cycles are
    forced in the diagonal-term calculation. This was found to be
    necessary in cases where asimov datasets were used for the
    minimization and there were very few iterations for the approximate
    covariance to be determined from.

  * Central finite difference is used for mixed partial derivatives.
    This requires 3 extra function evaluations per derivative, but is
    necessary in the case of minima where there is high curvature (in
    the case of high stats) and the forward finite difference (default)
    behaviour leads incorrectly to a non-positive-definite covariance
    matrix

  * Return the uncorrected covariance matrix, even if it is not positive
    definite. This useful for checking just how far from
    positive-definiteness the matrix is by being able to examine the
    eigenvalues.

Additionally, a lower bound on the precision allowed for the spread of
eigenvalues of the "hessian" correlation matrix (computing a correlation
matrix with the hessian as if it was a covariance matrix) was reduced
from 1e-6 to 1e-12 (see MnHesse.cxx) ... it is not clear why 1e-6 was
the lower bound previously, but current machine precision can beat that
(I get locally 1e-8). I left a comment about whether this lower bound
should be made configurable or not...

This new strategy was tested with a model with high statistics (almost
50 million events) where the migrad minimization was successful but the
hessian was being forced positive definite. With this new Strategy 3 the
hessian is accurate and positive definite in all cases tested.