[ntuple] Do not memset() a buffer that will be overwritten
`std::make_unique<unsigned char[]>()` has the side effect of zero-
initializing the allocated buffer.  In the replaced cases, all bytes
in the buffer are overwritten after the allocation.

Overhead of this initialization has been measured to be around 5-7%
in a tight loop that allocates and uses 100k 64KiB buffers.

This commit uses `std::unique_ptr<unsigned char[]>(new unsigned char[N])`
instead for the cases in which that overhead matters.
C++20 offers `std::make_unique_for_overwrite<T>()`, which might be used
in the future instead.