[ntuple] RClassField: pass both type (name|alias) to `RFieldBase::Create()`
Apparently, `TClassEdit::ResolveTypedef()` is unable to resolve a
typedef that aliases a fundamental data type if the `typedef` declaration
is not known to cling.  In non-cxxmodules builds, making it known
typically involves parsing the corresponding header.

This commit avoids that by directly using the resolved type name provided
by `TDataMember::GetTrueTypeName()`.