[RF] Don't store reference to `RooDataSet` in RooNDKeysPdf
It's not clear to users whether the `RooDataSet` passed to a
RooNDKeysPdf has to live as long as that pdf or not.

Right now it has to live just as long, because the RooNDKeysPdf contains
the RooDataSet by non-owning pointer. However, there was no reason for
that, because the dataset is only used in the constructor.

This commit suggests to drop this unused `_data` member of the
RooNDKeysPdf, solving potential lifetime problems: the dataset doesn't
need to live as long as the RooNDKeysPdf anymore.