[net] Improvements to `TPSocket` and `TPServerSocket` interfaces
Use the same default parameters for send and recieve functions as also
used in the non-parallel sockets, such that `kDefault` doesn't need to
be passed explicitly.

Also, change return type of `TPServerSocket::Accept()` to `TPSocket*`,
which is covariant with the return typoe `TSocket*` expected by the
virtual funciton `TServerSocket::Accept()`.