[clang] Backport patch to merge template member parent
Original message of upstream commit by Richard Smith,
https://github.com/llvm/llvm-project/commit/61c7a9140b:
---
Commit to a primary definition for a class when we load its first
member.

Previously, we wouldn't do this if the first member loaded is within a
definition that's added to a class via an update record, which happens
when template instantiation adds a class definition to a declaration
that was imported from an AST file.

This would lead to classes having member functions whose getParent
returned a class declaration that wasn't the primary definition, which
in turn caused the vtable builder to build broken vtables.

I don't yet have a reduced testcase for the wrong-code bug here, because
the setup required to get us into the broken state is very subtle, but
have confirmed that this fixes it.
---

This fixes an assertion in CodeGenFunction::EmitCXXDestructorCall():
Assertion `ThisTy->getAsCXXRecordDecl() == DtorDecl->getParent() &&
"Pointer/Object mixup"' failed.
which was already seen during the upgrade to LLVM 13 in one tutorial
on CentOS 8 and "solved" by commit ffe867985b ("Relax assertion on
generating destructor call"). Due to the nature of this problem, the
assertion failure went away with unrelated changes so I reverted the
change in 2b997add1d. Now the problem comes back with the upgrade to
LLVM 16 and also in master when trying to enable the RBatchGenerator
tutorials in https://github.com/root-project/root/pull/13302, both
on macOS this time. Luckily, the underlying cause was properly fixed
in upstream LLVM just last week, so backport that commit.