[RF] RooFitDriver refactoring and also take input arrays on the GPU
The `setData()` method of the RooFitDriver is replaced with a more
general `setInput()` method that can also take arrays on the GPU device.
Like this, one is not forced to have the input arrays on the host first
and then copy them back on the GPU. This will be useful for some
algorithms that want to evaluate RooFit objects on the GPU, but don't
take inputs from a RooDataSet but for example arrays that can be
generated directly on the GPU.

This commit also moves some functionality out of RooFitDriver.

This is to clean up the RooFitDriver interface to reflect purely a class
for evaluation on the CPU or GPU.

The plan is to eventually make the RooFitDriver public, such that users
have a clean way to evaluate RooFit PDFs on the GPU.