[math] Fix Pearson chi2 fit for weighted histograms
In case of a weighted histogram a weight was applied to compute the expected error for every histogram when the error is not equal to sqrt(N).  Now this is done only in case the user perform the fit using the new option "PW", otherwise it is assumed the histogram is not weighted and no correction is used for the expected error.